Add support for PNPM (#449)
* Revert "Remove all PNPM references"

This reverts commit f69071676cc18c3b3028abfa64c1194e8bb0edc8.

* A few others that needed to be reverted

* Add in PNPM stuff, resolve conflicts, rush generate

* Refactor linking strategy

* Remove debugging line

* Add Pete's comment

* PR feedback

* PR feedback

* PR Feedback and fixup the Shrinkwrap API

* Add more comments

* Add an example

* PR Feedback

* npm -> NPM and pnpm -> PNPM

* Fix a few places where we are referring to npm commands

+++ apps/rush-lib/src/cli/actions/GenerateAction.ts
@@ -7,6 +7,7 @@ import * as path from 'path';
 import * as fsx from 'fs-extra';
 import { CommandLineFlagParameter } from '@microsoft/ts-command-line';
 
+import { PackageManager } from '../../data/RushConfiguration';
 import Utilities from '../../utilities/Utilities';
 import { Stopwatch } from '../../utilities/Stopwatch';
 import InstallManager, { InstallType } from '../utilities/InstallManager';
@@ -23,6 +24,7 @@ export default class GenerateAction extends BaseRushAction {
   private _lazyParameter: CommandLineFlagParameter;
   private _noLinkParameter: CommandLineFlagParameter;
   private _forceParameter: CommandLineFlagParameter;
+  private _cleanParameter: CommandLineFlagParameter;
 
   constructor(parser: RushCommandLineParser) {
     super({
@@ -52,17 +54,37 @@ export default class GenerateAction extends BaseRushAction {
       parameterLongName: '--force',
       parameterShortName: '-f',
       description: 'Use this to bypass checking the shrinkwrap file, which forces rush generate to run even if all'
-      + ' dependencies already exist in the shrinkwrap file'
+      + ' dependencies already exist in the shrinkwrap file. Only applies when package manager is npm.'
     });
     this._noLinkParameter = this.defineFlagParameter({
       parameterLongName: '--no-link',
       description: 'Do not automatically run the "rush link" action after "rush generate"'
+    });
+    this._cleanParameter = this.defineFlagParameter({
+      parameterLongName: '--clean',
+      parameterShortName: '-c',
+      description: 'When using pnpm, forces a non-incremental clean install which clears the node_module and pnpm'
+        + ' store. Use this if any store corruption has occurred.'
     });
   }
 
   protected run(): void {
     const stopwatch: Stopwatch = Stopwatch.start();
     const isLazy: boolean = this._lazyParameter.value;
+
+    if (this._cleanParameter.value && this._lazyParameter.value) {
+      throw new Error(`Cannot specify both --clean and --lazy, as these are mutually exclusive operations.`);
+    }
+
+    if (this._lazyParameter.value && this.rushConfiguration.packageManager === 'pnpm') {
+      console.warn(colors.yellow('The --lazy flag is not required for pnpm'
+        + ' because its algorithm inherently incorporates this optimization.'));
+    }
+
+    if (this._cleanParameter.value && this.rushConfiguration.packageManager === 'npm') {
+      console.warn(colors.yellow('The --clean flag is not required for npm'
+        + ' because its algorithm always performs a clean installation.'));
+    }
 
     ApprovedPackagesChecker.rewriteConfigFiles(this.rushConfiguration);
 
@@ -104,47 +126,52 @@ export default class GenerateAction extends BaseRushAction {
       fsx.unlinkSync(tempShrinkwrapFilename);
     }
 
-    if (isLazy) {
-      console.log(colors.green(
-        `${os.EOL}Rush is running in "--lazy" mode. ` +
-        `You will need to run a normal "rush generate" before committing.`));
+    const packageManager: PackageManager = this.rushConfiguration.packageManager;
 
-      // Do an incremental install
-      installManager.installCommonModules(InstallType.Normal);
+    if (this.rushConfiguration.packageManager === 'pnpm') {
 
-      console.log(os.EOL + colors.bold('(Skipping "npm shrinkwrap")') + os.EOL);
-      const packageLogFilePath: string = path.join(this.rushConfiguration.commonTempFolder, 'package.lock');
+      // Do an incremental install unless --clean is specified
+      installManager.installCommonModules(this._cleanParameter.value
+        ?  InstallType.ForceClean
+        :  InstallType.Normal);
 
-      if (fsx.existsSync(packageLogFilePath)) {
-        console.log('Removing NPM5\'s "package.lock" file');
-        fsx.removeSync(packageLogFilePath);
-      }
-    } else {
-      // Do a clean install
-      installManager.installCommonModules(InstallType.ForceClean);
+      this._syncShrinkwrapAndCheckInstallFlag(installManager);
 
-      console.log(os.EOL + colors.bold('Running "npm shrinkwrap"...'));
-      const npmArgs: string[] = ['shrinkwrap'];
-      installManager.pushConfigurationArgs(npmArgs);
-      Utilities.executeCommand(this.rushConfiguration.packageManagerToolFilename,
-        npmArgs, this.rushConfiguration.commonTempFolder);
-      console.log('"npm shrinkwrap" completed' + os.EOL);
+    } else if (this.rushConfiguration.packageManager === 'npm') {
 
-      // Copy (or delete) common\temp\npm-shrinkwrap.json --> common\npm-shrinkwrap.json
-      installManager.syncFile(tempShrinkwrapFilename,
-        this.rushConfiguration.committedShrinkwrapFilename);
+      if (isLazy) {
+        console.log(colors.green(
+          `${os.EOL}Rush is running in "--lazy" mode. ` +
+          `You will need to run a normal "rush generate" before committing.`));
 
-      // The flag file is normally created by installCommonModules(), but "rush install" will
-      // compare its timestamp against the shrinkwrap file.  Since we just generated a new
-      // npm-shrinkwrap file, it's safe to bump the timestamp, which ensures that "rush install"
-      // won't do anything immediately after "rush generate".  This is a minor performance
-      // optimization, but it helps people to understand the semantics of the commands.
-      if (fsx.existsSync(installManager.commonNodeModulesMarkerFilename)) {
-        fsx.writeFileSync(installManager.commonNodeModulesMarkerFilename, '');
+        // Do an incremental install
+        installManager.installCommonModules(InstallType.Normal);
+
+        console.log(os.EOL + colors.bold('(Skipping "npm shrinkwrap")') + os.EOL);
+
+        // delete the automatically created npm5 "package.lock" file
+        const packageLogFilePath: string = path.join(this.rushConfiguration.commonTempFolder, 'package.lock');
+        if (fsx.existsSync(packageLogFilePath)) {
+          console.log('Removing NPM 5\'s "package.lock" file');
+          fsx.removeSync(packageLogFilePath);
+        }
       } else {
-        // Sanity check -- since we requested a clean install above, this should never occur
-        throw new Error('The install flag file is missing');
+        // Do a clean install
+        installManager.installCommonModules(InstallType.ForceClean);
+
+        console.log(os.EOL + colors.bold('Running "npm shrinkwrap"...'));
+        const npmArgs: string[] = ['shrinkwrap'];
+        installManager.pushConfigurationArgs(npmArgs);
+        Utilities.executeCommand(this.rushConfiguration.packageManagerToolFilename,
+          npmArgs, this.rushConfiguration.commonTempFolder);
+        console.log('"npm shrinkwrap" completed' + os.EOL);
+
+        this._syncShrinkwrapAndCheckInstallFlag(installManager);
       }
+
+    } else {
+      // program bug
+      throw new Error(`Program bug: invalid package manager "${packageManager}"`);
     }
 
     stopwatch.stop();
@@ -158,6 +185,24 @@ export default class GenerateAction extends BaseRushAction {
       this._parser.catchSyncErrors(linkManager.createSymlinksForProjects(true));
     } else {
       console.log(os.EOL + 'Next you should probably run: "rush link"');
+    }
+  }
+
+  private _syncShrinkwrapAndCheckInstallFlag(installManager: InstallManager): void {
+    // Copy (or delete) common\temp\npm-shrinkwrap.json --> common\npm-shrinkwrap.json
+    installManager.syncFile(this.rushConfiguration.tempShrinkwrapFilename,
+      this.rushConfiguration.committedShrinkwrapFilename);
+
+    // The flag file is normally created by installCommonModules(), but "rush install" will
+    // compare its timestamp against the shrinkwrap file.  Since we just generated a new
+    // npm-shrinkwrap file, it's safe to bump the timestamp, which ensures that "rush install"
+    // won't do anything immediately after "rush generate".  This is a minor performance
+    // optimization, but it helps people to understand the semantics of the commands.
+    if (fsx.existsSync(installManager.commonNodeModulesMarkerFilename)) {
+      fsx.writeFileSync(installManager.commonNodeModulesMarkerFilename, '');
+    } else {
+      // Sanity check -- since we requested a clean install above, this should never occur
+      throw new Error('The install flag file is missing');
     }
   }
 }
